work on pronouns profile command improvements
