--- bot/libs/ui/pronouns/__init__.py
@@ -1,7 +1,0 @@
-from .profile_pages import PronounsProfilePages as PronounsProfilePages
-from .structs import (
-    PronounsProfileCircleEntry as PronounsProfileCircleEntry,
-    PronounsProfileEntry as PronounsProfileEntry,
-    PronounsValuesEntry as PronounsValuesEntry,
-    PronounsWordsEntry as PronounsWordsEntry,
-)
