+++ requirements.txt
@@ -13,3 +13,4 @@ jishaku>=2.5.2,<3
 watchfiles>=0.21.0,<1
 PyYAML>=6.0.1,<7
 typing_extensions>=4.11.0,<5
+tabulate>=0.9.0,<1
