+++ taskfile.yml
@@ -1,0 +1,14 @@
+version: '3'
+
+tasks:
+  dev-up:
+    preconditions:
+      - test -f docker/docker-compose.dev.yml
+    cmds:
+      - docker compose -f docker/docker-compose.dev.yml up -d
+
+  dev-stop:
+    preconditions:
+      - test -f docker/docker-compose.dev.yml
+    cmds:
+      - docker compose -f docker/docker-compose.dev.yml stop
